requirements.txt
fastapi==0.115.4
uvicorn==0.30.6
sqlalchemy==2.0.36
pydantic-settings==2.5.2
argon2-cffi==23.1.0
PyJWT==2.9.0
python-multipart==0.0.9
structlog==24.4.0
stripe==9.5.0
cryptography==46.0.3
web3==6.19.0
eth-account==0.10.0
Pillow==11.0.0
imagehash==4.3.1
numpy==2.1.2
sentence-transformers==3.1.1
requests==2.31.0
